@@ -7,25 +7,34 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.google.android.material.tabs.TabLayoutMediator
-import com.myapplication.R
-import kotlinx.android.synthetic.main.fragment_tabs.view.*
+import com.myapplication.databinding.FragmentTabsBinding
 import timber.log.Timber
 
 
 class ItemTabsFragment : Fragment() {
 
+    private var _binding: FragmentTabsBinding? = null
+    // This property is only valid between `onCreateView` and `onDestroyView`
+    private val binding get() = _binding!!
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         Timber.d("onCreateView")
 
-        val rootView = inflater.inflate(R.layout.fragment_tabs, container,false)
+        // Inflate the layout for this fragment
+        _binding = FragmentTabsBinding.inflate(inflater, container, false)
 
-        rootView.viewpager.adapter = TabsPagerAdapter(this,3)
+        binding.viewpager.adapter = TabsPagerAdapter(this,3)
 
-        TabLayoutMediator(rootView.tab_layout, rootView.viewpager) { tab, pos ->
+        TabLayoutMediator(binding.tabLayout, binding.viewpager) { tab, pos ->
             tab.text = "TAB ${pos + 1}"
         }.attach()
 
-        return rootView
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
     inner class TabsPagerAdapter(fragment: Fragment, private val numTabs: Int): FragmentStateAdapter(fragment) {
