@@ -1,12 +1,12 @@
 package com.myapplication.activity
 
-import android.app.Application
 import androidx.annotation.Keep
+import androidx.multidex.MultiDexApplication
 import com.myapplication.BuildConfig
 import timber.log.Timber
 
 @Keep
-class MainApplication: Application() {
+class MainApplication: MultiDexApplication() {
 
     override fun onCreate() {
         super.onCreate()
