@@ -7,25 +7,33 @@ import android.view.ViewGroup
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
+import androidx.viewbinding.ViewBinding
 import com.myapplication.R
+import com.myapplication.databinding.FragmentHostSinglePaneBinding
+import com.myapplication.databinding.FragmentHostTwoPaneBinding
 import timber.log.Timber
 
 
 class MasterDetailHostFragment : Fragment() {
 
+    private var _binding: ViewBinding? = null
+    // This property is only valid between `onCreateView` and `onDestroyView`
+    private val binding get() = _binding!!
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         val isTablet = context?.resources?.getBoolean(R.bool.isTablet) ?: false
 
         Timber.d("onCreateView - isTablet: $isTablet")
 
-        val rootView = if (isTablet)
-            inflater.inflate(R.layout.fragment_host_two_pane, container, false)
+        // Inflate the layout for this fragment
+        _binding = if (isTablet)
+            FragmentHostTwoPaneBinding.inflate(inflater, container, false)
         else
-            inflater.inflate(R.layout.fragment_host_single_pane, container, false)
+            FragmentHostSinglePaneBinding.inflate(inflater, container, false)
 
         val master = childFragmentManager.findFragmentById(R.id.master_nav_fragment)
         master?.findNavController()?.navigate(R.id.master_fragment, arguments)
@@ -35,7 +43,12 @@ class MasterDetailHostFragment : Fragment() {
             detail?.findNavController()?.navigate(R.id.detail_fragment, arguments)
         }
 
-        return rootView
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
     }
 
 
